Fix uami timing and deploymentscript re-run issue (#39)
UAMI needs a few minutes to propagate through Azure AD and may cause timing issue in some tenants. So, a pause of 5 minutes is added after its creation.

The bicep template is not idempotent because the vnet and its subnets cannot be redeployed using bicep template. So, a solution was put in place earlier where deploymentScript would check the existence of the Vnet and conditionally deploy the vnet and its subnets. However, the deploymentScript does not re-run until the contents of the script changes.

So when it first runs, its output says the vnet does not exists. Subsequent deployments also use the same output since the deploymentScript is not re-run as the content of the script did not change.

As a fix, deleting the deploymentScript at the end of each run